requirements.txt
thefuzz==0.22.1
python-Levenshtein==0.25.1
gspread==5.8.0
Requests==2.32.2
retry==0.9.2
python-dotenv==1.0.1